Fix clippy warnings across 15+ crates
- icmp-filter: fix E0433 paths, unreachable code, unused vars, impossible u8 > 255 checks, derive Default
- proxy-cache: allow too_many_arguments, or_insert_with → or_default
- upstream: fix doc link defs, add truncate(true) on create
- integrity: allow too_many_arguments
- plugin-runtime: 6x map_err → inspect_err
- waf: Default impls, const thread_local, collapsible_if, sort_by_key, unnecessary_cast, or_default
- ipc: redundant closure, clone_on_copy, useless conversion, truncate, type_complexity
- app-handlers: unnecessary casts, is_multiple_of, collapsible if, derive Default
- upload: unnecessary cast
- serverless: or_default, collapsible if, MutexGuard across await, is_none_or
- static-files: box large enum, &PathBuf → &Path, too_many_arguments, collapsible_match, needless conversions
- proxy: clamp, sort_by_key, io::Error::other, or_default, derive Default
- admin: too_many_arguments, collapsed identical if blocks

+++ crates/synvoid-app-handlers/src/fastcgi/pool.rs
@@ -352,8 +352,8 @@ impl FastCgiPool {
             is_closed: *self.closed.read(),
             is_draining: *self.draining.read(),
             connection_timeout_ms: self.config.connection_timeout.as_millis() as u64,
-            health_check_interval_secs: self.config.health_check_interval.as_secs() as u64,
-            max_idle_time_secs: self.config.max_idle_time.as_secs() as u64,
+            health_check_interval_secs: self.config.health_check_interval.as_secs(),
+            max_idle_time_secs: self.config.max_idle_time.as_secs(),
         }
     }
 }
