+++ crates/synvoid-waf/src/ratelimit/sliding.rs
@@ -177,7 +177,7 @@ impl<K: Hash + Eq + Clone> SlidingWindowLimiter<K> {
     pub fn new(configs: Vec<SlidingWindowConfig>, max_entries: usize) -> Self {
         Self {
             entries: parking_lot::RwLock::new(std::collections::HashMap::with_capacity_and_hasher(
-                max_entries.next_power_of_two().max(64) as usize,
+                max_entries.next_power_of_two().max(64),
                 std::hash::RandomState::default(),
             )),
             configs,
